fix: review findings on the pashov-audit fixes
Drawdown floor scaled off the wrong quantity. `maxDrawdownBps` is declared,
validated and documented as a share of COMMITTED capital, but the gate computed
`vaultBalance * (1 - bps)`, so a proposal committing 10% of the float was
allowed to lose ten times its own declared envelope before tripping. Measure the
absolute drop against `effectiveMaxCapital * bps` instead — the same
coverage-scaled figure execute and the settlement batch are bounded by. As a
side effect every balance component that is constant across the window (the
queue reserve, an untouched fee escrow) now cancels exactly instead of scaling
the bar.

Escrowed fees could leave the vault mid-strategy. `claimUnclaimedFees` ->
`transferPerformanceFee` was the one asset outflow between execute and settle
that is neither the strategy nor gated on `redemptionsLocked()`. Both consumers
of the vault's balance difference read it as a strategy loss: it understates
`_finishSettlement`'s pnl, and since the drawdown floor it can revert an
otherwise-profitable settle — which leaves `_activeProposal` set and locks
redemptions, queue claims and every future proposal until the owner runs
`unstick`. Gate the claim on the claimed vault's active proposal; the escrow
accrues no deadline, so the recipient loses nothing but a wait.

Guardian review probe could not degrade as documented. `abi.decode(ret,
(uint64, uint64))` REVERTS on a word whose high bits exceed the narrow type, and
that revert lands in the caller's frame with no `try` around it — a registry
returning a malformed window would have bricked `_commitState` and every
`getProposalState` read rather than falling back to Expired. Decode wide; the
only question asked of the word is `!= 0`.

Also documents what `_tryQuoteMinOut` does NOT soften: an unavailable quote is
non-fatal, an overstated one still reverts the swap it floors, and in Data
Streams mode that now reaches `settle()`.

Tests: the envelope-vs-whole-fund discriminator (the fraction case, where the
two formulas diverge), the escrow-claim gate, and `unstick` clearing a proposal
the floor refused — the escape hatch the gate's scoping argument rests on.

+++ src/ProposalLifecycle.sol
@@ -189,11 +189,19 @@ abstract contract ProposalLifecycle is ISyndicateGovernor {
             // "no window", i.e. the pre-existing terminal-Expired behaviour:
             // fail-closed, and the vault binding is released rather than held
             // to `executeBy` on the strength of a reply we could not read.
+            //
+            // DECODED AS `uint256`, NOT `uint64`, even though `reviewWindow`
+            // declares `uint64`. `abi.decode` REVERTS on a word whose high bits
+            // exceed the narrow type, and that revert lands in THIS frame with
+            // no `try` around it — it would brick `_commitState` and every
+            // `getProposalState` read rather than degrading. Decoding wide
+            // cannot revert on any 64-byte payload, and the only question asked
+            // of the word is `!= 0`, which is scale-independent.
             bool windowRegistered;
             (bool probeOk, bytes memory ret) =
                 address(reg).staticcall(abi.encodeCall(IGuardianRegistry.reviewWindow, (address(this), p.id)));
             if (probeOk && ret.length >= 64) {
-                (, uint64 registeredReviewEnd) = abi.decode(ret, (uint64, uint64));
+                (, uint256 registeredReviewEnd) = abi.decode(ret, (uint256, uint256));
                 windowRegistered = registeredReviewEnd != 0;
             }
             // (a) no record: terminal immediately, so `_commitState` releases the
